// src/enum/timeline-actions.enum.ts
export enum TIMELINE_ACTIONS {
  LOGIN = "LOGIN",
  LOGOUT = "LOGOUT",
  CHANGE_PASSWORD = "CHANGE_PASSWORD",
  EMAIL = "EMAIL",
  CREATE = "CREATE", // REGISTER= 'REGISTER',
  UPDATE = "UPDATE",
  DELETE = "DELETE",
  PAYMENT = "PAYMENT",
  REGISTER_ORDER = "REGISTER_ORDER",
  ORDER_RECEIVE = "ORDER_RECEIVE",
  ORDER_RETURN = "ORDER_RETURN",
  COUPON_USED = "COUPON_USED",
}